+++ testTypes/Balance.test.ts
@@ -4,7 +4,7 @@ import starkbank from "starkbank";
 import assert from 'assert';
 
 starkbank.user = require('./utils/user').exampleProject;
-starkbank.Balance
+
 describe('TestBalanceGet', function(){
     it('test_success', async () => {
         let balance = await starkbank.balance.get();